Refactor Storage from io package to util package
