+++ src/main/java/baymax/util/Parser.java
@@ -5,7 +5,6 @@ import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 import baymax.exception.BaymaxException;
-import baymax.io.Storage;
 import baymax.task.Deadline;
 import baymax.task.Event;
 import baymax.task.Task;
