+++ src/main/java/baymax/util/Storage.java
@@ -1,4 +1,4 @@
-package baymax.io;
+package baymax.util;
 
 
 import java.io.File;
